Prettier (#12)
* format via prettier

* prettier fixed

+++ contracts/libraries/LibDiamondStorage.sol
@@ -2,35 +2,36 @@
 pragma solidity ^0.7.1;
 
 library LibDiamondStorage {
-
-    bytes32 constant DIAMOND_STORAGE_POSITION = keccak256("diamond.standard.diamond.storage");
+    bytes32 constant DIAMOND_STORAGE_POSITION = keccak256(
+        "diamond.standard.diamond.storage"
+    );
 
     struct DiamondStorage {
-
         // owner of the contract
         address contractOwner;
-
         // maps function selectors to the facets that execute the functions.
         // and maps the selectors to the slot in the selectorSlots array.
         // and maps the selectors to the position in the slot.
         // func selector => address facet, uint32 slotIndex, uint64 slotsIndex
         mapping(bytes4 => bytes32) facets;
-
         // array of slots of function selectors.
         // each slot holds 8 function selectors.
-        mapping(uint => bytes32) selectorSlots;
-
+        mapping(uint256 => bytes32) selectorSlots;
         // The number of function selectors in selectorSlots
-        uint selectorCount;
-
+        uint256 selectorCount;
         // Used to query if a contract implements an interface.
         // Used to implement ERC-165.
         mapping(bytes4 => bool) supportedInterfaces;
     }
 
-
-    function diamondStorage() internal pure returns(DiamondStorage storage ds) {
+    function diamondStorage()
+        internal
+        pure
+        returns (DiamondStorage storage ds)
+    {
         bytes32 position = DIAMOND_STORAGE_POSITION;
-        assembly { ds.slot := position }
+        assembly {
+            ds.slot := position
+        }
     }
 }
